feat: Extract out legacy os exporter and unify StandaloneSchemaManager
